Expose rigid contact kinematics at root
Match simulation evaluation APIs such as eval_fk by exporting the contact helper as newton.eval_rigid_contact_kinematics.

Keep the caller-provided output contract while moving the public wrapper into the simulation layer.
