+++ docs/api/newton_geometry.rst
@@ -39,7 +39,6 @@ newton.geometry
    collide_sphere_sphere
    compute_inertia_shape
    compute_offset_mesh
-   compute_rigid_contact_kinematics
    create_empty_sdf_data
    sdf_box
    sdf_capsule
