+++ newton/__init__.py
@@ -95,6 +95,7 @@ from ._src.sim import (  # noqa: E402
     eval_inverse_dynamics_passive,
     eval_jacobian,
     eval_mass_matrix,
+    eval_rigid_contact_kinematics,
 )
 
 __all__ += [
@@ -116,6 +117,7 @@ __all__ += [
     "eval_inverse_dynamics_passive",
     "eval_jacobian",
     "eval_mass_matrix",
+    "eval_rigid_contact_kinematics",
 ]
 
 # ==================================================================================
