+++ newton/_src/sim/__init__.py
@@ -4,6 +4,7 @@
 from .articulation import eval_fk, eval_ik, eval_inverse_dynamics_force, eval_jacobian, eval_mass_matrix
 from .builder import ModelBuilder
 from .collide import CollisionPipeline
+from .contact_kinematics import eval_rigid_contact_kinematics
 from .contacts import Contacts
 from .control import Control
 from .enums import (
@@ -37,4 +38,5 @@ __all__ = [
     "eval_inverse_dynamics_passive",
     "eval_jacobian",
     "eval_mass_matrix",
+    "eval_rigid_contact_kinematics",
 ]
